Preview: Improve accessibility of loading screen
