+++ .changeset/thirty-cities-rest.md
@@ -1,0 +1,5 @@
+---
+'playroom': patch
+---
+
+Preview: Improve accessibility of loading screen
